move validation to separate hook and reuse it for DateRangeForm
