+++ graylog2-web-interface/src/views/components/time-range-picker/RelativeRangeSelect.tsx
@@ -23,10 +23,9 @@ import moment from 'moment';
 import { RELATIVE_RANGE_TYPES } from 'views/Constants';
 import { Select } from 'components/common';
 
-import type { RangeClassified } from './types';
 import { isTypeRelativeClassified, RELATIVE_CLASSIFIED_ALL_TIME_RANGE } from './RelativeTimeRangeClassifiedHelper';
 import RelativeRangeValueInput from './RelativeRangeValueInput';
-import type { TimeRangePickerFormValues } from './TimeRangePicker';
+import type { RangeClassified, TimeRangePickerFormValues } from './types';
 
 const RangeWrapper = styled.div`
   flex: 4;
@@ -191,6 +190,7 @@ const RelativeRangeSelectInner = ({
       <RangeTitle>{title}</RangeTitle>
       <RangeCheck htmlFor={`relative-unset-${fieldName}`} className={disableUnsetRange && 'shortened'}>
         <input
+          aria-label={unsetRangeLabel}
           checked={isAllTime}
           className="mousetrap"
           disabled={disableUnsetRange}
