+++ graylog2-web-interface/src/views/components/time-range-picker/TimeRangePresetRow.tsx
@@ -26,7 +26,7 @@ import { IfPermitted } from 'components/common';
 import SaveTimeRangeAsPresetButton from 'views/components/time-range-picker/SaveTimeRangeAsPresetButton';
 import TimeRangeInputSettingsContext from 'views/components/contexts/TimeRangeInputSettingsContext';
 import type { TimeRange } from 'views/logic/queries/Query';
-import type { TimeRangePickerFormValues } from 'views/components/time-range-picker/TimeRangePicker';
+import type { TimeRangePickerFormValues } from 'views/components/time-range-picker/types';
 import { classifyRelativeTimeRange } from 'views/components/time-range-picker/RelativeTimeRangeClassifiedHelper';
 import { ButtonToolbar } from 'components/bootstrap';
